rename: 트리 등록페이지 컴포넌트 디렉토리명 변경
