+++ src/components/RegistTree/LocationSearch/SearchResult/index.tsx
@@ -3,7 +3,7 @@ import Item from 'components/common/Item';
 import { useDispatch, useSelector } from 'react-redux';
 import ISearchPlaceData from 'types/SearchPlaceData';
 import usePlaceSearch from 'hooks/usePlaceSearch';
-import { setAddressType, setLatitude, setLongitude } from 'store/modules/locationSlice';
+import { setAddressType, setLatitude, setLongitude } from 'store/modules/treeRegistLocationSlice';
 import { Link } from 'react-router-dom';
 import * as S from './style';
 
@@ -22,6 +22,7 @@ function SearchTip() {
 
 function SearchListItem({ data }: { data: ISearchPlaceData }) {
   const dispatch = useDispatch();
+  const addressName = data.road_address_name || data.address_name;
 
   const setRegistLocation = () => {
     const addressType = data.road_address_name ? 'ROAD' : 'STREET';
@@ -39,7 +40,7 @@ function SearchListItem({ data }: { data: ISearchPlaceData }) {
         </Item.Title>
         <S.ResultItemAddress>
           <S.AddressType>{data.road_address_name ? '도로명' : '지번'}</S.AddressType>
-          {data.road_address_name || data.address_name}
+          {addressName}
         </S.ResultItemAddress>
       </Link>
     </S.ResultListItem>
