+++ src/components/RegistTree/LocationSearch/SearchResult/style.ts
